Dispatch publishers per feature
